Initial task refactoring
This performs the initial refactoring of the application in order
to make it possible to extend the code to support downloading and
installing OPDS packages.

The existing task monad types were replaced with a more principled
type, and the tasks themselves were replaced with implementations
that aggressively retry operations in the case of failures.

Additionally, a service directory type was introduced in order to
reduce coupling between application components.