Update the inventory/repository models to handle OPDS
This updates the code to subscribe to the OPDS database for events
that indicate whether or not an OPDS catalog is "installed". This
adds events to the OPDS database that are published when
catalogs are deleted, and renames the "InstalledItems" API to
"InstalledApplications" to make it clearer that it only deals with
Android applications. This also refactors event publication in the
InventoryRepository* classes to avoid publishing events with locks held
(because this risks deadlocks).

Affects: https://github.com/AULFA/updater/issues/12