Add a boot screen
This adds a more principled boot screen and uses a service
configuration similar to the SimplyE app. This allows for more work
to be done in the app on startup without resulting in a long blank
screen on startup. It also adjusts how bundled repositories are
instantiated to ensure easy upgrades.