Ensure consistent OPDS version code generation
This adjusts the procedure used to derive version codes from OPDS
manifests. We essentially convert the input date to UTC (which may
be an identity operation), and then format the date as a timestamp
without punctuation, and convert that directly to an integer value.

Fix: https://github.com/AULFA/updater/issues/22