Introduce installation passwords
This introduces a new v3 format that allows APKs and OPDS feeds to be
"protected" with hashed passwords. That is, the updater will prompt for
a password and the hash of the password must match that of the entry
in the repository. This first commit just updates the data structures
with the new information, and introduces a parser and serializer for
the v3 format. The updater does not yet prompt for a password.